pancancer-328 continued
This allows youxia to deploy on-demand when an exception occurs on the
amazon side for spot instances. It also fixes reaper behaviour when
SimpleDB is not available (Frankfurt) (also disables proper handling of
sensu though).

+++ youxia-reaper/src/main/java/io/cloudbindle/youxia/reaper/Reaper.java
@@ -377,23 +377,25 @@ public class Reaper {
     }
 
     private void mergePersistentRecord(Map<String, String> instancesToKill) {
-        AmazonSimpleDBClient simpleDBClient = ConfigTools.getSimpleDBClient();
-        final String query = "select * from `" + deletedClientsDomain + "`" + " where " + dayOfTheYearString + " = \"" + dayOfTheYear
-                + "\" and " + yearString + " = \"" + year + "\"";
-        // get information on previously cleaned instances for today and merge with today's
-        createDomainIfRequired(simpleDBClient, deletedClientsDomain);
-        SelectRequest select = new SelectRequest(query, true);
-        SelectResult selectResult = simpleDBClient.select(select);
-        for (Item item : selectResult.getItems()) {
-            handleAttribute(item, instancesToKill);
-        }
-        while (selectResult.getNextToken() != null) {
-            selectResult = simpleDBClient.select(new SelectRequest(query, true).withNextToken(select.getNextToken()));
+        if (options.has(this.persistWR)) {
+            AmazonSimpleDBClient simpleDBClient = ConfigTools.getSimpleDBClient();
+            final String query = "select * from `" + deletedClientsDomain + "`" + " where " + dayOfTheYearString + " = \"" + dayOfTheYear
+                    + "\" and " + yearString + " = \"" + year + "\"";
+            // get information on previously cleaned instances for today and merge with today's
+            createDomainIfRequired(simpleDBClient, deletedClientsDomain);
+            SelectRequest select = new SelectRequest(query, true);
+            SelectResult selectResult = simpleDBClient.select(select);
             for (Item item : selectResult.getItems()) {
                 handleAttribute(item, instancesToKill);
             }
+            while (selectResult.getNextToken() != null) {
+                selectResult = simpleDBClient.select(new SelectRequest(query, true).withNextToken(select.getNextToken()));
+                for (Item item : selectResult.getItems()) {
+                    handleAttribute(item, instancesToKill);
+                }
+            }
+            Log.stdoutWithTime("Merged kill list is " + StringUtils.join(instancesToKill, ','));
         }
-        Log.stdoutWithTime("Merged kill list is " + StringUtils.join(instancesToKill, ','));
     }
 
     private void handleAttribute(Item item, Map<String, String> instancesToKill) {
@@ -440,6 +442,7 @@ public class Reaper {
                 reaper.terminateInstances(instancesToKill);
             }
         }
+
         reaper.mergePersistentRecord(instancesToKill);
         reaper.terminateSensuClients(test, Sets.newHashSet(instancesToKill.values()));
     }
